-AssetbundleV2：Refactorying  build assetbundle node logic.
