+++ DevOps/PublishAssets/Android/PackageBuild.Info
@@ -1,1 +1,1 @@
-{"BuildTime":1648835479,"AssetBundleSVCVersion":"","ScriptSVCVersion":"","TableSVCVersion":""}
+{"BuildTime":1648841066,"AssetBundleSVCVersion":"","ScriptSVCVersion":"","TableSVCVersion":""}
