+++ DevOps/PublishAssets/Android/ServerAssetsSubPackage.Conf
@@ -1,3 +1,4 @@
 PackageName,ArtAssetsIdList,HotfixCodePathList,TablePathList,ConfAndInfoList
-TestChar,"[-1]","[Hotfix/hotfix.dll]","[Local.db]","[Art/ArtConfig.Info,Art/AssetTypeConfig.Info,PackageBuild.Info]"
+TestChar,"[31,50,51,52,59]","[Hotfix/hotfix.dll]","[Local.db]","[Art/ArtConfig.Info,Art/AssetTypeConfig.Info,PackageBuild.Info]"
 TestNull,"[]","[Hotfix/hotfix.dll]","[Local.db]","[Art/ArtConfig.Info,Art/AssetTypeConfig.Info,PackageBuild.Info]"
+TestWindow,"[19,33,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,58]","[Hotfix/hotfix.dll]","[Local.db]","[Art/ArtConfig.Info,Art/AssetTypeConfig.Info,PackageBuild.Info]"
