+++ Packages/com.popo.bdframework/Editor/EditorPipeline/BuildPipeline/AssetBundleEditor/AssetBundleEditorToolsV2CheckAssetbundle.cs
@@ -138,15 +138,15 @@ namespace BDFramework.Editor.AssetBundle
 
             foreach (var asset in allRuntimeAssets)
             {
-                var type = AssetBundleBuildToolsV2.GetMainAssetTypeAtPath(asset);
+                var type = AssetBundleBuildingContext.GetMainAssetTypeAtPath(asset);
                 if (type == null)
                 {
                     Debug.LogError("无法获得资源类型:" + asset);
                     continue;
                 }
 
-                var idx = asset.IndexOf(AssetBundleBuildToolsV2.RUNTIME_PATH, StringComparison.OrdinalIgnoreCase);
-                var runtimePath = asset.Substring(idx + AssetBundleBuildToolsV2.RUNTIME_PATH.Length);
+                var idx = asset.IndexOf(AssetBundleBuildingContext.RUNTIME_PATH, StringComparison.OrdinalIgnoreCase);
+                var runtimePath = asset.Substring(idx + AssetBundleBuildingContext.RUNTIME_PATH.Length);
                 runtimePath = runtimePath.Replace(Path.GetExtension(runtimePath), "");
                 runtimePath = runtimePath.Replace("\\", "/");
                 //Debug.Log("【LoadTest】:" + runtimePath);
