--- Packages/com.popo.bdframework/Editor/EditorPipeline/BuildPipeline/AssetBundleEditor/AssetGraph/Core/IBDFrameowrkAssetEnvParams.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 398fb7cc2f614f1e8479de724fc45c10
-timeCreated: 1632505025
