+++ Packages/com.popo.bdframework/Editor/EditorPipeline/BuildPipeline/AssetBundleEditor/AssetGraph/Node/SetGranularityByFolder.cs
@@ -15,10 +15,12 @@ namespace BDFramework.Editor.AssetGraph.Node
     /// 颗粒度,排序30-50
     /// </summary>
     [CustomNode("BDFramework/[颗粒度]文件夹规则", 30)]
-    public class SetGranularityByFolder : UnityEngine.AssetGraph.Node, IBDFrameowrkAssetEnvParams
+    public class SetGranularityByFolder : UnityEngine.AssetGraph.Node
     {
-        public BuildAssetsInfo BuildAssetsInfo { get; set; }
-        public BuildAssetBundleParams BuildParams { get; set; }
+        /// <summary>
+        /// 构建的上下文信息
+        /// </summary>
+        public AssetBundleBuildingContext BuildingCtx { get; set; }
 
         public void Reset()
         {
@@ -140,10 +142,10 @@ namespace BDFramework.Editor.AssetGraph.Node
             {
                 return;
             }
+            this.BuildingCtx = BDFrameworkAssetsEnv.BuildingCtx;
 
             Debug.Log("prepare:" + this.GetType().Name + "-" + DateTime.Now.ToLongTimeString());
-            this.BuildAssetsInfo = BDFrameworkAssetsEnv.BuildAssetsInfo;
-            this.BuildParams = BDFrameworkAssetsEnv.BuildParams;
+
 
 
             //
@@ -186,7 +188,7 @@ namespace BDFramework.Editor.AssetGraph.Node
                     foreach (var ar in ag.Value)
                     {
                         //设置当前ab名为文件夹名,不覆盖在此之前的规则
-                        var ret = BuildAssetsInfo.SetABName(ar.importFrom, folderPath);
+                        var ret = this.BuildingCtx.BuildAssetsInfo.SetABName(ar.importFrom, folderPath);
                         if (!ret)
                         {
                             Debug.LogError($"【颗粒度】设置AB失败 [{folderPath}] -" + ar.importFrom);
@@ -248,7 +250,7 @@ namespace BDFramework.Editor.AssetGraph.Node
                         {
                             if (ar.importFrom.StartsWith(sf + "/", StringComparison.OrdinalIgnoreCase))
                             {
-                                var ret = BuildAssetsInfo.SetABName(ar.importFrom, sf);
+                                var ret =  this.BuildingCtx.BuildAssetsInfo.SetABName(ar.importFrom, sf);
                                 if (!ret)
                                 {
                                     Debug.LogError($"【颗粒度】设置AB失败 [{sf}] -" + ar.importFrom);
@@ -265,7 +267,7 @@ namespace BDFramework.Editor.AssetGraph.Node
                         if (!isInSubfolder)
                         {
                             //设置AB name
-                            var ret = BuildAssetsInfo.SetABName(ar.importFrom, rootfloderPath);
+                            var ret =  this.BuildingCtx.BuildAssetsInfo.SetABName(ar.importFrom, rootfloderPath);
                             if (!ret)
                             {
                                 Debug.LogError($"【颗粒度】设置AB失败 [{rootfloderPath}] -" + ar.importFrom);
