+++ Packages/com.popo.bdframework/Editor/EditorPipeline/BuildPipeline/AssetBundleEditor/EditorWindow_GenAssetBundle.cs
@@ -173,7 +173,7 @@ namespace BDFramework.Editor.AssetBundle
             }
             
             //生成Assetbundlebunle
-            AssetBundleBuildToolsV2.GenAssetBundle(outputPath, platform);
+            AssetBundleBuildingContext.GenAssetBundle(outputPath, platform);
             AssetDatabase.Refresh();
             Debug.Log("资源打包完毕");
         }
