+++ Packages/com.popo.bdframework/Editor/EditorPipeline/PublishPipeline/EditorBehavior/ABDFrameworkPublishPipelineBehaviour.cs
@@ -66,16 +66,15 @@ namespace BDFramework.Editor
         /// <summary>
         /// 一键打包时，开始导出AssetBundle
         /// </summary>
-        /// <param name="buildAssetsInfo">自定义修改buildinfo内容 进行自定义的ab输出</param>
-        virtual public void OnBeginBuildAssetBundle(BuildAssetBundleParams @params, BuildAssetsInfo buildAssetsInfo)
+      
+        virtual public void OnBeginBuildAssetBundle(AssetBundleBuildingContext assetbundleBuildingCtx)
         {
         }
 
         /// <summary>
         ///  一键打包时，完成导出AssetBundle
         /// </summary>
-        /// <param name="outputPath">dll输出路径</param>
-        virtual public void OnEndBuildAssetBundle(BuildAssetBundleParams @params, BuildAssetsInfo buildAssetsInfo)
+        virtual public void OnEndBuildAssetBundle(AssetBundleBuildingContext assetbundleBuildingCtx)
         {
         }
